Move plugin-specific code into plugins, revert gui SonarCloud changes
- Move SilentSaxErrorHandler from core into xml-format and xml-xsd-validator plugins
- Revert AutoFormatPayload and test SonarCloud refactoring (no gui changes)
- Make SilentSaxErrorHandler package-private in each plugin
