+++ gui/src/test/java/org/correomqtt/PublishViewControllerFormattingTest.java
@@ -214,7 +214,6 @@ class PublishViewControllerFormattingTest extends ApplicationTest {
 
         @Override
         public void setText(String text) {
-            // Intentionally empty: test stub that ignores setText calls
         }
 
         @Override
