+++ plugins/xml-format/src/main/java/org/correomqtt/plugin/xml_format/SilentSaxErrorHandler.java
@@ -1,0 +1,23 @@
+package org.correomqtt.plugin.xml_format;
+
+import org.xml.sax.ErrorHandler;
+import org.xml.sax.SAXException;
+import org.xml.sax.SAXParseException;
+
+final class SilentSaxErrorHandler implements ErrorHandler {
+
+    @Override
+    public void warning(SAXParseException exception) {
+        // Intentionally empty: warnings are non-fatal and safely ignored during XML validation
+    }
+
+    @Override
+    public void error(SAXParseException exception) throws SAXException {
+        throw exception;
+    }
+
+    @Override
+    public void fatalError(SAXParseException exception) throws SAXException {
+        throw exception;
+    }
+}
